main/requirements.txt
pandas
numpy
pillow
pillow-heif
opencv-python
pytesseract